chore: rename LogPayment component to LogPaymentModal
