+++ src/components/invoices/invoice-list.tsx
@@ -1,20 +1,20 @@
 import { InvoiceStatusBadge } from "@atoms/badge";
+import Button from "@atoms/button";
 import Loading from "@atoms/loading";
 import ListPage from "@components/shared/list-page";
+import { faPlus } from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { InvoiceStatus } from "@prisma/client";
 import { InvoiceListOutput } from "@server/api/routers/invoice-router";
+import { getTotalCostOfActivities } from "@utils/activity-utils";
 import { trpc } from "@utils/trpc";
 import classNames from "classnames";
+import Link from "next/link";
 import { useState } from "react";
-import { getTotalCostOfActivities } from "@utils/activity-utils";
+import LogPayment from "./log-payment-modal";
 
 import dayjs from "dayjs";
 import utc from "dayjs/plugin/utc";
-import Button from "@atoms/button";
-import LogPayment from "./log-payment";
-import Link from "next/link";
-import { faPlus } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 dayjs.extend(utc);
 
 interface Props {
